Added support for SNP
Old way should be unaltered by editing srp-index instead of srindex,
from now on it's possible to link based on a simplified channelname. If
an exception should be made, it's still possible to add the old way to
the file snp-index. Enigma2 tuners will in this case prefer the
serviceref of the channelname.